Implement specific flags for auto saving and restoring and update docs
New flags should start as nil, unless the default behaviour of the
plugin is not the desired behaviour.

Setting the `auto_save_enabled = true` for example allows for a session
to be saved even when vim was first opened as a "pager". i.e passing in
parameters for reading from stdin.

TODO: fix the vim portion not calling auto save or auto restore if in
pager mode, rely on lua configs instead of VimL pager mode setting.

Closes #13.