Friendlier command completion Show the actual unescaped 'name' of the session instead of the full path to the session + escaped name when showing completion options.
Allow :DeleteSession to take more than one argument. This requires
the underlying lua functions (DeleteSessionByName and DeleteSession) to
take variable args also.

If the user uses :DeleteSession <arg> to delete the current session the
session management is turned off. This is the same functionality as
when :DeleteSession is used with no arguments.