Add DirChangedPre and DirChanged handling
This commit adds handling for when the cwd changes by setting callbacks
to both `DirChangedPre` and `DirChanged` autocmds.

Handling is as follows
  DirChangedPre:
    - Save the current session
    - Clear all buffers `%bd!`. This guarantees buffers don't bleed to the
      next session.
    - Clear jumps. Also done so there is no bleading between sessions.
    - Run the `pre_cwd_changed_hook`
  DirChanged:
    - Restore session using new cwd
    - Run the `post_cwd_changed_hook`

Now when the user changes the cwd with `:cd some/new/dir` auto-session
handles it gracefully, saving the current session so there aren't losses and loading the session for the upcoming cwd if it
exists.

Closes #102.