Make the cwd_change_handling feature be disabled by default.
Many users had issues where they were using some plugins that changed
the cwd frequently and would run into an undesired situation with the
new feature. This commit makes it so the default behaviour is whatever
auto session used to do before implementing this feature so as to not
break other people's workflows.

Fixes #157 and #158