add AutoSession.session_exists_for_cwd() function
Add a function to understand if a session exists for the given current
working directory.

This is pretty helpful for start page plugins like mini.starter and
others.

By calling this function when nvim is opened to a workspace folder the
starter page can determine whether to show a 'restore session' option or
not.

Signed-off-by: ldelossa <louis.delos@gmail.com>