Check for Lazy window on VimEnter, added config
Changed the code to check for the Lazy window on the VimEnter event
(there's no benefit to waiting for the VeryLazy event and it would slow
down normal loading when Lazy doesn't pop up a window). If the window is
present, we wait until we get a close event for the Lazy window (as
before)

Added a config option: auto_restore_lazy_delay_enabled. It's not
necessary to change as Lazy is autodetected but it might be useful for
debugging or narrowing down issues if someone is having trouble with
autorestore. Setting it to false with Lazy installed will produce the
pre-fix behavior of restoring the session into the Lazy window