Fix tried to close last window error
Easiest repro is only have a new buffer (not saved to a file) and exit.
We'll  try to close that window and generate an error.

To fix, we wrap `Lib.close_unsupported_windows` in a `pcall`. In
general, not great to eat all errors but we're making our best effort to
close all of the unsupported windows and there are a bunch of edge cases
there so this seems like the most reasonable path to me.

Also added a missiing local