Fix for #299
Fix the conf checks in session-lens/actions.lua to handle the case
where cwd_change_handling is true.

The original bug report includes a config for when session lens was
in a separate repo and that confused me for a bit so I tried to make
the documentation more clear.

Also, added the Telescope load_extension call in setup_session_lens so
users don't have to.

Lastly, I think there's a bug with the built-in picker (:Autosession
search) where it won't load the session if cwd_change_handling is false.
I did not address that in this PR but left a BUG: marker.