Unify session selection logic, Fixes #287
Unify the session loading logic between Telesecop session-lens and
AutoSession search so that we always use the session file to load the
session (vs relying on changing directory). Set a flag to make sure
cwd_change_handling (if enabled) doesn't also try to restore the session
This also fixes a session load failure when a session with a git branch
name is selected from Telescope session-lens or Autosession search.

Removed the non-functional (as far as I can see) code for
buftypes_to_ignore. That code would not bwipeout a buffer if it matched
but the buffers get wiped out by the sourcing the session file anyway
and having that code generates an error when the buffer is opened in
another process which interrupts the session swap. Made the option
deprecated and asked anyone using it to report it on GH.