Fix #295 by detecting if *x.vim is a session file
The session check works as follows:
For each session file, first check that it's not a directory.
Then check if it ends in x.vim. If it doesn't, then return true.
If it does, if the first line contains the word
"SessionLoad" then return true and false otherwise.

This will include valid session files that happen to end in x and not
include the *x.vim session files neovim creates for custom user commands
saved with the session.