fix: #200 Prevent cd loading a session while already restoring sesssion
If `cwd_change_handling = true` and `sessionoptions` contains `curdir`,
the session will contain a cd command to change to the directory where
the session was saved. That would trigger DirChangedPre/DirChanged and
DirChanged would try to load the session again. Obviously, we don't want
to do that. We have previously prevented it for AutoRestoreSession and
loading a session via search but it's better to do it in RestoreSession
to handle all of the cases.