fix: shellslash support in legacy files
I don't think shellslash=true ever worked because it never escaped the
'/'. The legacy encoding isn't relevant if you don't have old sessions
(which I don't think could be generated with shellslash on) so shouldn't
matter with urlencoding going forward but I want to fix it, if nothing
else, so the unit tests pass.