fix: #335 session_root_dir didn't have a path separator
One of the rare places where I got the root dir without the separator.
Since we already expanded the root dir when validating, we already know
it'll be absolute so don't need to use plenary to get an absolute path
later. So we can get the root dir as normal. Only showed up for *x.vim
sessions because that's the only time we need to actually open the file
to make sure it's not a user commands file.