build: make running tests faster
Specify that the tests should load the minimal.lua config. It's a
bit confusing, but calling `nvim -u mininmal.lua` loads the minimal
config for starting the Plenary test harness. The test harness then
starts another, nvim subprocess to actually run the test. Specifying
`minimal_init` tells the test harness to not load other plugins (the
minimal part) and to load the init file we pass in (which we'll use
later to load Telescope for one of the tests)

Also, run the test via PlenaryBustedDirectory instead of PlenaryBustedFile.
This is a bit faster because it only starts the test harness process
once rather than starting the harness for every spec file.

Also make the args tests not require a setup so each file is independent
and move them to the main directory.