fix: can't delete a session from Telescope picker
Sometimes deleting a session from the the Telescope picker wasn't
working. It happens when we have extra command files that are filtered
out from the list. Those files break some code in Telescope that finds
the item we want to delete in the picker results. More detail is here:

https://github.com/nvim-telescope/telescope.nvim/issues/3265

We work around the issue by having the Telescope entry maker returning
{} instead of nil. That's not perfect but better than not being able to
delete/