fix: #358 Don't bubble up ignored errors
We had been relying on `v:errmsg` to catch session errors when restoring
with `silent! source <session>`. The problem with that approach is
`v:errmsg` might be set even when it's an error message that should be
ignored (e.g. by another `silent!` command).

Instead, we now `silent source <session>` (note no !) but if that fails
(which means it's a real error), then we optionally source the session
again, this time with silent! so we load as much of the session as
possible.