refactor(Config)!: Clean up picker customization
Rename `theme_conf` to `picker_opts` to better represent what it does.

Deprecated `shorten_path` (even harder this time) and `path_display`
when under the `session_lens` config block. Both of those are Telescope
specific and should be in `picker_opts`. We'll automatically detect when
they're still in `session_lens` and set them in `picker_opts` and show
the old config warning `checkhealth`.

Added support for configuring the session picker when using Snacks via
`picker_opts`.

Change is only listed as breaking because the `previewer` option was
removed. The default is still that no preview is displayed (for both
Telescope and Snacks) but if you want to turn preview on for the session
picker for some reason, set the appropriate field in `picker_opts`
(`previewer = true,` for Telescope and `preview = true,` for Snacks)